feat: отделен модул за showLoading
